+++ node/resolvers/catalog/recommendationsResolver.ts
@@ -1,17 +1,12 @@
-import axios from 'axios'
-import {map, prop} from 'ramda'
-import paths from '../paths'
-import {resolveLocalProductFields} from './fieldsResolver'
-import {withAuthToken} from '../headers'
+import axios from 'axios';
+import { map } from 'ramda';
 
-export const resolveView = async (ioContext, product) => {
-  const url = paths.crossSelling(ioContext.account, product.productId, 'whosawalsosaw')
-  const {data} = await axios.get(url, { headers: withAuthToken()(ioContext) })
-  return map(resolveLocalProductFields, data)
-}
+import { withAuthToken } from '../headers';
+import paths from '../paths';
+import { resolveLocalProductFields } from './fieldsResolver';
 
-export const resolveBuy = async (ioContext, product) => {
-  const url = paths.crossSelling(ioContext.account, product.productId, 'whoboughtalsobought')
+export const resolveRecommendation = async (ioContext, product, type) => {
+  const url = paths.crossSelling(ioContext.account, product.productId, type)
   const {data} = await axios.get(url, { headers: withAuthToken()(ioContext) })
   return map(resolveLocalProductFields, data)
 }