+++ node/resolvers/headers.ts
@@ -1,5 +1,7 @@
 import * as cookies from 'cookie'
 
+const DEFAULT_PAGE_SIZE = 15
+
 export const headers = {
   json: {
     accept: 'application/json',
@@ -12,8 +14,7 @@ export const headers = {
 }
 
 export const withAuthToken = (currentHeaders = {}) => (ioContext, cookie = null) => {
-  let VtexIdclientAutCookie
-  let ans = {...currentHeaders}
+  let ans = { ...currentHeaders }
   if (cookie) {
     const parsedCookie = cookies.parse(cookie)
     ans['VtexIdclientAutCookie'] = parsedCookie.VtexIdclientAutCookie
@@ -22,5 +23,25 @@ export const withAuthToken = (currentHeaders = {}) => (ioContext, cookie = null)
     ...ans,
     Authorization: `${ioContext.authToken}`,
     'Proxy-Authorization': `${ioContext.authToken}`
+  }
+}
+
+export const withAuthAsVTEXID = (currentHeaders = {}) => (ioContext) => {
+  return {
+    ...currentHeaders,
+    VtexIdclientAutCookie: ioContext.authToken,
+    'Proxy-Authorization': ioContext.authToken,
+  }
+}
+
+export const withMDPagination = (currentHeaders = {}) => (ioContext, cookie = null) => (page = 0, pageSize = DEFAULT_PAGE_SIZE) => {
+  if (page < 1) {
+    throw new Error('Smallest page value is 1')
+  }
+  const startIndex = (page - 1) * pageSize
+  const endIndex = startIndex + pageSize
+  return {
+    ...(withAuthToken(currentHeaders)(ioContext, cookie)),
+    'REST-Range': `resources=${startIndex}-${endIndex}`
   }
 }